Add convenience method `g` on `Labrador`
This commit adds a convenience method `g(uri)` on the `Labrador`
object.  This method retrieves the value only, omitting status
information.  Previously this was only avaiable on the module-level
`get()` convenience.  Given this new function, this commit also makes
the module-level convenience function expressed in terms of the
object-level function.